pyproject.toml
[project]
name = "ai-pdf-analizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.131.0",
    "inngest>=0.5.15",
    "llama-index-core>=0.14.15",
    "llama-index-readers-file>=0.5.6",
    "ollama>=0.6.1",
    "openai>=2.21.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.17.0",
    "streamlit>=1.54.0",
    "uvicorn>=0.41.0",
]
